Break up the geom.go file, resolves #5
