+++ handlers/geospatialdata.go
@@ -3,7 +3,7 @@ package handlers
 import (
 	"net/http"
 
-	"github.com/USACE/mcat-ras/config"
+	"github.com/USACE/filestore"
 	ras "github.com/USACE/mcat-ras/tools"
 
 	"github.com/labstack/echo/v4"
@@ -19,17 +19,17 @@ import (
 // @Success 200 {object} interface{}
 // @Failure 500 {object} SimpleResponse
 // @Router /geospatialdata [get]
-func GeospatialData(cf *config.APIConfig) echo.HandlerFunc {
+func GeospatialData(fs *filestore.FileStore, destinationCRS int) echo.HandlerFunc {
 	return func(c echo.Context) error {
 
 		definitionFile := c.QueryParam("definition_file")
 
-		rm, err := ras.NewRasModel(definitionFile, *cf.FileStore)
+		rm, err := ras.NewRasModel(definitionFile, *fs)
 		if err != nil {
 			return c.JSON(http.StatusInternalServerError, SimpleResponse{http.StatusInternalServerError, err.Error()})
 		}
 
-		data, err := rm.GeospatialData(cf.DestinationCRS)
+		data, err := rm.GeospatialData(destinationCRS)
 		if err != nil {
 			return c.JSON(http.StatusInternalServerError, SimpleResponse{http.StatusInternalServerError, err.Error()})
 		}
